Add files via upload
An extended solver to account for scalar subgrid turbulent diffusion calculation. The procedure is based on the work of Lilly (1992) that employed a dynamic approach similar to the dynamic Smagorinsky method for turbulent viscosity.
The solver must be used along with SDSM which involves the dynamic Smagorinsky model, but with the scalar diffusion part added.